// web-frontend/norbitz/src/apis/hurts-car-rental/model/vehicle.ts
/**
 * Hurts Car API
 * Hurts Car Rental API
 *
 * OpenAPI spec version: 1.1.1b
 * Contact: support@softwarebois.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { SpecialEquipment } from './specialEquipment';


export interface Vehicle {
    vehicleID: number;
    quantity: number;
    make: string;
    model: string;
    year: number;
    location: string;
    cost: number;
    passengers: number;
    autoTransmission: boolean;
    type: string;
    mpg: number;
    specialEquipment: SpecialEquipment;
}