// web-frontend/norbitz/src/apis/hurts-car-rental/model/specialEquipment.ts
/**
 * Hurts Car API
 * Hurts Car Rental API
 *
 * OpenAPI spec version: 1.1.1b
 * Contact: support@softwarebois.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface SpecialEquipment {
    gps: boolean;
    maxChildSeat: number;
    skiRack: boolean;
    snowChains: boolean;
    leftControl: boolean;
}